requirements.txt
fastapi~=0.104.1
pydantic[email]~=2.5.2
requests~=2.31.0
uvicorn~=0.24.0
psycopg2-binary~=2.9.9
passlib[bcrypt]~=1.7.4
bcrypt~=3.2.2
python-jose[cryptography]~=3.3.0
python-multipart~=0.0.6